+++ src/main/java/net/sf/jabref/JabRefPreferences.java
@@ -102,6 +102,9 @@ public class JabRefPreferences {
     public static final String PROXY_PORT = "proxyPort";
     public static final String PROXY_HOSTNAME = "proxyHostname";
     public static final String USE_PROXY = "useProxy";
+    public static final String PROXY_USERNAME = "proxyUsername";
+    public static final String PROXY_PASSWORD = "proxyPassword";
+    public static final String USE_PROXY_AUTHENTICATION = "useProxyAuthentication";
     public static final String TABLE_PRIMARY_SORT_FIELD = "priSort";
     public static final String TABLE_PRIMARY_SORT_DESCENDING = "priDescending";
     public static final String TABLE_SECONDARY_SORT_FIELD = "secSort";
@@ -454,8 +457,11 @@ public class JabRefPreferences {
         }
         defaults.put(PUSH_TO_APPLICATION, "TeXstudio");
         defaults.put(USE_PROXY, Boolean.FALSE);
-        defaults.put(PROXY_HOSTNAME, "my proxy host");
-        defaults.put(PROXY_PORT, "my proxy port");
+        defaults.put(PROXY_HOSTNAME, "");
+        defaults.put(PROXY_PORT, "80");
+        defaults.put(USE_PROXY_AUTHENTICATION, Boolean.FALSE);
+        defaults.put(PROXY_USERNAME, "");
+        defaults.put(PROXY_PASSWORD, "");
         defaults.put(PDF_PREVIEW, Boolean.FALSE);
         defaults.put(USE_DEFAULT_LOOK_AND_FEEL, Boolean.TRUE);
         defaults.put(LYXPIPE, System.getProperty("user.home") + File.separator + ".lyx/lyxpipe");